// Boardlight/src/api/generated/models/shopIndexOwnOwnedItemsResponse.ts
/**
 * Generated by orval v6.16.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */
import type { ShopIndexOwnOwnedItemsResponseProduct } from './shopIndexOwnOwnedItemsResponseProduct';

export interface ShopIndexOwnOwnedItemsResponse {
  id?: number;
  product?: ShopIndexOwnOwnedItemsResponseProduct;
  usedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}