// Boardlight/src/api/generated/models/shopUseOwnedItemRequestBody.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */
import type { ShopUseOwnedItemRequestBodyInputs } from './shopUseOwnedItemRequestBodyInputs';

export interface ShopUseOwnedItemRequestBody {
  qrCodeContent?: string | null;
  inputs?: ShopUseOwnedItemRequestBodyInputs;
}